refactor(v0.6.1): split model_resolver + intent_classifier under rule #5 cap
Resolves the two pre-existing rule-#5 violations grandfathered by the
vision-wire PR (both were RED before that PR — model_resolver.py via the
routing build-out, intent_classifier.py via #987 + vision registration):

- core/model_resolver_tiers.py (NEW) ← Phase-3a LOCAL_TIER_LADDER +
  _local_tier_tag + resolve_local_tier. One-way import from
  model_resolver (no cycle). model_resolver.py 21,042 → 17,097 B.
- core/intent_fast_patterns.py (NEW) ← the classify_fast FAST_PATTERNS
  regex table. IntentClassifier.FAST_PATTERNS now references the module
  dict (byte-identical). intent_classifier.py 21,067 → 12,203 B.

Only caller of the moved tier symbols (scripts/research/
routing_matrix_probe.py) updated to import from the new module. No other
importers. classify_fast parity verified (chat/meta/coding/wiki_edit).

GRANDFATHERED emptied per the anti-creep rule — every core/**/*.py is
back at/under 20 KB. Module-size gate GREEN forward-only again (no
grandfather escape hatch).

core/retrieval + core/graph traversal: 0 lines. core/reasoning: 0 lines
(this touches core/model_resolver + core/intent_classifier only, both
structural moves — behaviour byte-identical).
